+++ backend/src/neuroagent/agent_routine.py
@@ -150,7 +150,22 @@ class AgentsRoutine:
                 }
                 return response, None
 
-        tool_metadata = tool.__annotations__["metadata"](**context_variables)
+        try:
+            tool_metadata = tool.__annotations__["metadata"](**context_variables)
+        except ValidationError as err:
+            # Raise validation error if requested
+            if raise_validation_errors:
+                raise err
+            else:
+                # Otherwise transforn it into an OpenAI response for the model to retry
+                response = {
+                    "role": "tool",
+                    "tool_call_id": tool_call.tool_call_id,
+                    "tool_name": name,
+                    "content": "The user is not allowed to run this tool. Don't call it again.",
+                }
+                return response, None
+
         tool_instance = tool(input_schema=input_schema, metadata=tool_metadata)
         # pass context_variables to agent functions
         try: