+++ frontend/src/components/chat/chat-message-ai.tsx
@@ -1,40 +1,31 @@
-import Markdown from "react-markdown";
-import remarkGfm from "remark-gfm";
 import { Card, CardContent } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
 import { LoaderPinwheel, ChevronDown, Wrench } from "lucide-react";
-
-function isSubset<T>(smallSet: Set<T>, largeSet: Set<T>): boolean {
-  return [...smallSet].every((element) => largeSet.has(element));
-}
+import { MemoizedMarkdown } from "@/components/memoized-markdown";
 
 type ChatMessageAIProps = {
-  id: string;
   content?: string;
-  threadId: string;
-  associatedToolsIncides: string[];
-  collapsedTools: Set<string>;
-  toggleCollapse: (messageId: string[]) => void;
+  hasTools: boolean;
+  toolsCollapsed: boolean;
+  toggleCollapse: () => void;
+  messageId: string;
 };
 
-export function ChatMessageAI({
+export const ChatMessageAI = function ChatMessageAI({
   content,
-  associatedToolsIncides,
-  collapsedTools,
+  hasTools,
+  toolsCollapsed,
   toggleCollapse,
+  messageId,
 }: ChatMessageAIProps) {
   return (
     <div className="flex justify-start mt-4">
-      {associatedToolsIncides.length > 0 ? (
+      {hasTools ? (
         <Button
           className="hover:scale-105 active:scale-[1.10] ml-7 mr-1 mt-0.5 bg-blue-500 rounded-full p-2.5"
-          onClick={() => toggleCollapse(associatedToolsIncides)}
+          onClick={toggleCollapse}
         >
-          {isSubset(new Set(associatedToolsIncides), collapsedTools) ? (
-            <Wrench />
-          ) : (
-            <ChevronDown />
-          )}
+          {toolsCollapsed ? <Wrench /> : <ChevronDown />}
         </Button>
       ) : (
         <Button className="ml-7 mr-1 mt-0.5 bg-blue-500 rounded-full p-2.5 hover:bg-blue-500">
@@ -45,10 +36,10 @@ export function ChatMessageAI({
       <Card className="max-w-[70%] bg-transparent shadow-none border-none mt-1">
         <CardContent>
           <span className="prose text-lg pt-8 text-left dark:prose-invert">
-            <Markdown remarkPlugins={[remarkGfm]}>{content}</Markdown>
+            <MemoizedMarkdown content={content || ""} id={messageId} />
           </span>
         </CardContent>
       </Card>
     </div>
   );
-}
+};