+++ frontend/src/lib/cookies-client.ts
@@ -1,8 +1,0 @@
-import Cookies from "js-cookie";
-
-export function getSettings() {
-  return {
-    projectID: Cookies.get("projectID") ?? "",
-    virtualLabID: Cookies.get("virtualLabID") ?? "",
-  };
-}