+++ CHANGELOG.md
@@ -57,8 +57,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Displaying tool metadata
 - GET threads depends on vlab+project_id.
 - Object storage + plotting logic.
+- Keywords filtering for literature search.
 - [frontend] - bugfixes.
 - Mutli-agent workflow.
+- Suggestion endpoint.
+- Support empty vlab+project in threads.
 
 ### Fixed
 - [frontend] - Parallel execution of tools
@@ -82,6 +85,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [frontend] - Handle invalid json errors.
 - [frontend] - Tool selection is all checked when refreshing.
 - Fix bug for AI + tool calls.
+- [frontend] - Fix docker compose redirect bug.
+- [frontend] - Refactor
+- Websearch tools.
 
 ## [0.3.3] - 30.10.2024